+++ src/app/component/trash-time-off/trash-time-off.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { TimeOffService } from 'src/app/service/time-off.service';
-
+import { ToastsService } from 'src/app/service/toasts.service';
+import Swal from 'sweetalert2';
 @Component({
   selector: 'app-trash-time-off',
   templateUrl: './trash-time-off.component.html',
@@ -8,7 +9,7 @@ import { TimeOffService } from 'src/app/service/time-off.service';
 })
 export class TrashTimeOffComponent implements OnInit {
   listTimeOffDelete = []
-  constructor(private timeOffService:TimeOffService) { }
+  constructor(private timeOffService:TimeOffService,private toastService:ToastsService) { }
 
   ngOnInit(): void {
     this.getAllDeleteTimeOff()
@@ -16,20 +17,75 @@ export class TrashTimeOffComponent implements OnInit {
   getAllDeleteTimeOff(){
     this.timeOffService.getAllDelete().subscribe((data)=> {
      
-      this.listTimeOffDelete = data
+      this.listTimeOffDelete = data.data
       console.log(this.listTimeOffDelete);
     })
   }
   handleDestroy(id:string){
-    this.timeOffService.destroyTimeOff(id).subscribe((data)=> {
-      console.log(data);
-      
+    const swalWithBootstrapButtons = Swal.mixin({
+      customClass: {
+        confirmButton: 'btn btn-success',
+        cancelButton: 'btn btn-danger'
+      },
+      buttonsStyling: false
     })
-  }
+    swalWithBootstrapButtons.fire({
+      title: 'Bạn chắc chắn muốn xóa vĩnh viễn',
+      text: "Không thể khôi phục sau khi xóa !",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Có',
+      cancelButtonText: 'Không',
+      reverseButtons: true
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.timeOffService.destroyTimeOff(id).subscribe((data)=> {
+          console.log(data);
+          this.toastService.show(data.message,{
+            classname:'bg-success text-light',
+            delay:3000
+          }),
+          (err:any)=>{
+            this.toastService.show(err.message,{
+              classname:'bg-danger text-light',
+              delay:3000
+            })
+          }
+        })
+        this.getAllDeleteTimeOff()
+      }else if(result.dismiss === Swal.DismissReason.cancel){
+        this.getAllDeleteTimeOff()
+      }
+   
+  });
+}
   handleRestore(id:string,object:any){
-    this.timeOffService.restoreTimeOff(id,object).subscribe((data)=> {
-      console.log(data);
-      
+    const swalWithBootstrapButtons = Swal.mixin({
+      customClass: {
+        confirmButton: 'btn btn-success',
+        cancelButton: 'btn btn-danger'
+      },
+      buttonsStyling: false
     })
+    swalWithBootstrapButtons.fire({
+      title: 'Bạn chắc chắn chứ',
+      text: "Khôi phục dữ liệu",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Có',
+      cancelButtonText: 'Không',
+      reverseButtons: true
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.timeOffService.restoreTimeOff(id,object).subscribe((data)=> {
+          console.log(data);
+          
+        })
+        this.getAllDeleteTimeOff()
+      }else if(result.dismiss === Swal.DismissReason.cancel){
+        this.getAllDeleteTimeOff()
+      }
+   
+  });
   }
 }
